% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{add_style_to_styles}
\alias{add_style_to_styles}
\title{Add style_node to docx_styles as a child node}
\usage{
add_style_to_styles(docx_styles, style_node)
}
\arguments{
\item{docx_styles}{\code{docx_styles} object. (both \code{xml_node} and \code{xml_document} are OK)}

\item{style_node}{\code{xml_node} of style.}
}
\value{
\code{docx_styles} updated by argument style_node as its child.
}
\description{
Add style_node to docx_styles as a child node
}
