% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{copy_style_from_docx_styles}
\alias{copy_style_from_docx_styles}
\title{Make a new style copied from referenced style in docx_styles}
\usage{
copy_style_from_docx_styles(
  docx_styles,
  new_style_name,
  ref_style_name = NULL,
  style_id = "auto"
)
}
\arguments{
\item{docx_styles}{\code{docx_styles} object with \code{styles} at its root.}

\item{new_style_name}{new style name as a \code{character} vector.}

\item{ref_style_name}{referenced style name in the \code{xml} as a \code{character} vector. Default is \code{NULL}. It means \code{"Normal"} style.}

\item{style_id}{new style id. Unique id in docx is required. By default, it is set by unique id.}
}
\value{
new style as a \code{xml_node} object.
}
\description{
Make a new style copied from referenced style in docx_styles
}
