% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{copy_style_from_node}
\alias{copy_style_from_node}
\title{Make a style node from reference node}
\usage{
copy_style_from_node(style_node, style_name)
}
\arguments{
\item{style_node}{reference style node of \code{xml_node}}

\item{style_name}{name for new style}
}
\value{
new style node of \code{xml_node}
}
\description{
Make a style node from reference node
}
