% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{create_docx_styles_root}
\alias{create_docx_styles_root}
\title{Create root of \code{docx_styles} object.}
\usage{
create_docx_styles_root()
}
\value{
\code{docx_styles} object which has \code{styles} as a root.
}
\description{
Create root of \code{docx_styles} object.
}
