% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{create_new_style}
\alias{create_new_style}
\title{Create new style from reference style}
\usage{
create_new_style(df, style_name, ref_style_name = NULL, style_id = "auto")
}
\arguments{
\item{df}{style data.frame}

\item{style_name}{style name}

\item{ref_style_name}{reference style's name}

\item{style_id}{new style id.}
}
\value{
new style by column?
}
\description{
Create new style from reference style
}
