% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{create_styles_root}
\alias{create_styles_root}
\title{Internal function for making xml_node with style at the top}
\usage{
create_styles_root()
}
\value{
xml_node with styles root.
}
\description{
Internal function for making xml_node with style at the top
}
