% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.calculate_content_loss}
\alias{.calculate_content_loss}
\title{Calculate Content Loss}
\usage{
.calculate_content_loss(content_layer, generated_layer)
}
\arguments{
\item{content_layer}{Features extracted from a specific layer of the content
image.}

\item{generated_layer}{Features extracted from the same layer of the
generated image. The tensor must be the same shape as the content image. A
common cause for error is the content image and the generated image having
different height and/or width.}
}
\value{
A length-1 object inheriting from class "torch_tensor" with the loss
between the generated layer and the content layer.
}
\description{
Content loss is calculated by comparing tensor output from the target content
layer to the tensor output from the generated layer.
}
\keyword{internal}
