% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.calculate_style_layer_loss}
\alias{.calculate_style_layer_loss}
\title{Calculate Style Layer Loss}
\usage{
.calculate_style_layer_loss(style_layer, generated_layer)
}
\arguments{
\item{style_layer}{Features extracted from a single style layer.}

\item{generated_layer}{Features extracted from a single generated layer.
Should be the same layer as style_layer.}
}
\value{
A scalar torch_tensor representing the loss for this layer.
}
\description{
The style loss for each layer is calculated by comparing the Gram matrix for
that style layer to the Gram matrix for the equivalent generated layer. The
intention is for this to act as a measurement that the different channels
that make up a layer are activated to the same proportion between the style
layer and the generated layer. For example, if this layer "cares" that all
vertical lines in the style layer are orange, measure how untrue that is for
the generated layer.
}
\keyword{internal}
