% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.calculate_style_loss}
\alias{.calculate_style_loss}
\title{Calculate Style Loss}
\usage{
.calculate_style_loss(
  input_style_features,
  generated_style_features,
  style_lambdas
)
}
\arguments{
\item{input_style_features}{A list of tensors, each containing features
extracted from a single layer for which style matters.}

\item{generated_style_features}{A list of tensors, the same length as
input_style_features, with the equivalent features from the generated
image.}

\item{style_lambdas}{The weights associated with each layer.}
}
\value{
A length-1 object inheriting from class "torch_tensor" with the loss
between the generated layers and the style layers.
}
\description{
Style loss is calculated by weighting loss from multiple layers. For each
chosen layer, the loss measures the relative activation of the channels
making up that layer.
}
\keyword{internal}
