% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate_gram_matrix}
\alias{.generate_gram_matrix}
\title{Generate Gram Matrix}
\usage{
.generate_gram_matrix(tensor)
}
\arguments{
\item{tensor}{A 1 x C x H x W tensor.}
}
\value{
A tensor with shape C x C.
}
\description{
Gram matrices can be used to determine how much different channels are
dependent on one another. This dependence is used as a definition of "style."
}
\keyword{internal}
