% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.weight_style_layer_loss}
\alias{.weight_style_layer_loss}
\title{Calculate the Weighted Style Layer Loss}
\usage{
.weight_style_layer_loss(style_layer, generated_layer, lambda)
}
\arguments{
\item{style_layer}{Features extracted from a single style layer.}

\item{generated_layer}{Features extracted from a single generated layer.
Should be the same layer as style_layer.}

\item{lambda}{The weight to apply to this loss.}
}
\value{
A scalar torch_tensor representing the weighted loss for this layer.
}
\description{
A simple convenience function to multiple loss by weight.
}
\keyword{internal}
