% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nodes.R
\name{ext_parent_tag}
\alias{ext_parent_tag}
\title{Existence check of parent of tag}
\usage{
ext_parent_tag(node, tag)
}
\arguments{
\item{node}{xml node/node_set object of style}

\item{tag}{tag under the node to be checked}
}
\value{
TRUE : if parent tag exist or no parent tag.
FALSE : if parent tag does not exist.
}
\description{
Existence check of parent of tag
}
