% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_attr_from_node}
\alias{get_attr_from_node}
\title{Get attribution value from specified node}
\usage{
get_attr_from_node(node, attr, ctag = NULL)
}
\arguments{
\item{node}{style tag as \code{xml_node}}

\item{attr}{attribution name as a character vector.}

\item{ctag}{child tag name as a character if needed.}
}
\value{
value of attribution
}
\description{
Get attribution value from specified node
}
