% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeutils.R
\name{get_id_by_name}
\alias{get_id_by_name}
\title{Get style id from style name}
\usage{
get_id_by_name(styles_xml, name)
}
\arguments{
\item{styles_xml}{xml2 node object of styles.}

\item{name}{style name to get style_id.}
}
\value{
style_id
}
\description{
Get style id from style name
}
\examples{
\dontrun{
 get_id_by_name(xml, "Author")
}
}
