% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeutils.R
\name{get_name_by_id}
\alias{get_name_by_id}
\title{Get style name from style id}
\usage{
get_name_by_id(style_xml, style_id)
}
\arguments{
\item{style_xml}{xml2 node object of styles.}

\item{style_id}{style_id to get style name.}
}
\value{
style name
}
\description{
Get style name from style id
}
\examples{
\dontrun{
 get_name_by_id(xml, "Author")
}
}
