% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeutils.R
\name{get_node_by_id}
\alias{get_node_by_id}
\title{Utility functions realted to xml nodes.}
\usage{
get_node_by_id(style_xml, node_id)
}
\arguments{
\item{node_id}{}
}
\value{

}
\description{
To get specific node by name, id and so on.
}
\examples{


}
