% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeutils.R
\name{get_node_by_name}
\alias{get_node_by_name}
\title{To get specific node by style name}
\usage{
get_node_by_name(styles_xml, name)
}
\arguments{
\item{styles_xml}{xml2 node object of styles.}

\item{name}{style name to get specific node.}
}
\value{
xml2 node object
}
\description{
To get specific node by style name
}
\examples{
\dontrun{
 get_node_by_name(xml, "Author")
}
}
