% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{get_style_ids}
\alias{get_style_ids}
\title{style id vector}
\usage{
get_style_ids(df)
}
\arguments{
\item{df}{styles data.frame}
}
\value{
character vector of style id in df.
}
\description{
style id vector
}
