% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeutils.R
\name{get_style_tags_from_styles}
\alias{get_style_tags_from_styles}
\title{Get style tags of xml_nodeset under style tag of xml_node object}
\usage{
get_style_tags_from_styles(styles_xml)
}
\arguments{
\item{styles_xml}{xml_node object}
}
\value{
xml_nodeset object which contains style tags
}
\description{
Get style tags of xml_nodeset under style tag of xml_node object
}
