% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_styles}
\alias{get_styles}
\title{Extract only style tags and its children}
\usage{
get_styles(xml)
}
\arguments{
\item{xml}{expects xml2 node object which contains tags from styles.xml such as output of \code{read_style_xml}.}
}
\value{
xml2 node object which contains style tags.
}
\description{
Extract only style tags and its children
}
\examples{
\dontrun{
get_styles(xml)
}
}
