% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeutils.R
\name{is_same_nodes}
\alias{is_same_nodes}
\title{Check nodes are same except their id and name}
\usage{
is_same_nodes(node1, node2)
}
\arguments{
\item{node1}{xml_node which contains 1 style}

\item{node2}{xml_node which contains 1 style}
}
\value{
True if same. Others False.
}
\description{
Check nodes are same except their id and name
}
