% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{new_style_from_styles}
\alias{new_style_from_styles}
\title{Create a new style to reference style based on existing style}
\usage{
new_style_from_styles(
  styles_xml,
  style_name,
  ref_name = NULL,
  style_id = "auto"
)
}
\arguments{
\item{styles_xml}{\code{xml_node} object with \code{styles} at the root.}

\item{style_name}{new style name as a \code{character} vector.}

\item{ref_name}{referenced style name in the \code{xml} as a \code{character} vector. Default is \code{NULL}. It means \code{"Normal"} style.}

\item{style_id}{new style id. Unique id in docx is required. By default, it has unique id.}
}
\value{
new style as a \code{xml_node} object.
}
\description{
Create a new style to reference style based on existing style
}
