% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_style_xml}
\alias{read_style_xml}
\title{Extract styles.xml file in the docx file}
\usage{
read_style_xml(file)
}
\arguments{
\item{file}{Input file path of reference document file (docx).}
}
\value{
node object of xml2 which contains all tags in styles.xml file.
}
\description{
Extract styles.xml file in the docx file
}
\examples{

\dontrun{
  read_style_xml("reference.docx")
}

}
