% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_styles}
\alias{read_styles}
\title{Extract styles.xml file in the docx file}
\usage{
read_styles(file)
}
\arguments{
\item{file}{Input file path of reference document file (docx).}
}
\value{
\code{node_document} object which contains all tags in styles.xml file.
}
\description{
Extract styles.xml file in the docx file
}
\examples{

\dontrun{
  read_docx("reference.docx")
}

}
