% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nodes.R
\name{set_attr_val}
\alias{set_attr_val}
\title{Common function to update attr by val}
\usage{
set_attr_val(node, val, tag, attr)
}
\arguments{
\item{node}{xml2 node style object.}

\item{val}{value for update.}

\item{tag}{tag under the node.}

\item{attr}{attribute name with "w:" prefix.}
}
\value{
None
}
\description{
Common function to update attr by val
}
