% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{set_based_on_val}
\alias{set_based_on_val}
\title{internal functions to modify node's value by val}
\usage{
set_based_on_val(node, val)
}
\arguments{
\item{val}{}
}
\value{

}
\description{
internal functions to modify node's value by val
}
