% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nodes.R
\name{set_exist_tag}
\alias{set_exist_tag}
\title{set existence to target tag}
\usage{
set_exist_tag(node, val, ptag, ctag)
}
\arguments{
\item{node}{xml single style node}

\item{val}{set value. non-NA or NA.}

\item{ptag}{parent tag}

\item{ctag}{child tag}
}
\value{
xml style node object after modification
}
\description{
set existence to target tag
}
