% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nodes.R
\name{set_exist_tag_wo_ctag}
\alias{set_exist_tag_wo_ctag}
\title{set existence to target tag w/o child tag.}
\usage{
set_exist_tag_wo_ctag(node, val, tag)
}
\arguments{
\item{node}{xml single style node}

\item{val}{set value. NA or non-NA.}

\item{tag}{target tag}
}
\value{
updated single xml style tag.
}
\description{
set existence to target tag w/o child tag.
}
