% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nodes.R
\name{set_node_tagattr}
\alias{set_node_tagattr}
\title{Set value to tag/attr of givnen node}
\usage{
set_node_tagattr(node, val, tag, attr, flag = NULL)
}
\arguments{
\item{node}{\code{xml_node} or \code{xml_nodeset}}

\item{val}{value to set to the node}

\item{tag}{tag name as a character}

\item{attr}{attribute name as a character}

\item{flag}{TBD.}
}
\value{
updated node (node in arguments is automatically updated)
}
\description{
Set value to tag/attr of givnen node
}
