% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2df.R
\name{style2dddf}
\alias{style2dddf}
\title{Style_xml to data.frame for docDefaults part.}
\usage{
style2dddf(style_xml)
}
\arguments{
\item{style_xml}{xml_nodeset of docx which contains only style tags.}
}
\value{
data.frame contains docDefaults parts.
}
\description{
\code{docDefaults} part in the docx file are separated from main function.
}
