% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2df.R
\name{style2df}
\alias{style2df}
\title{Convert xml object to data.frame}
\usage{
style2df(style_xml)
}
\arguments{
\item{style_xml}{A node oject of xml2. It expexts output of \code{read_style_xml()}.}
}
\value{
data.frame converted from input xml.
}
\description{
Convertion from xml to data.frame for modification usability.
}
\examples{
\dontrun{
style2df(style_xml)
}

}
