% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdf2xml.R
\name{update_xml}
\alias{update_xml}
\title{Update xml file based on reference.docx and data.frame.}
\usage{
update_xml(docx_file, df)
}
\arguments{
\item{docx_file}{reference.docx file which has original styles.}

\item{df}{data.frame which contains updated value. Note that all updated values are updated.}
}
\value{
Updated style xml node object.
}
\description{
Update xml file based on reference.docx and data.frame.
}
\examples{
\dontrun{
update_xml("reference.docx", df)
}

}
