% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{write_docx}
\alias{write_docx}
\title{Write docx file with original reference.docx and updated style xml}
\usage{
write_docx(style_xml, org_docx, new_docx)
}
\arguments{
\item{style_xml}{updated style xml node object.}

\item{org_docx}{original reference.docx file name as a character.}

\item{new_docx}{updated docx file name as a character.}
}
\value{
None
}
\description{
Write docx file with original reference.docx and updated style xml
}
\examples{
\dontrun{
  write_docx(xml, "reference.docx", "updated.docx")
}

}
