% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{write_styles}
\alias{write_styles}
\title{Write docx file with original reference.docx and updated style xml}
\usage{
write_styles(styles_xml, ref_file, new_file)
}
\arguments{
\item{styles_xml}{\code{xml_node} object with \code{styles}}

\item{ref_file}{reference docx file name as a \code{character}. Use files except \code{word/styles.xml}.}

\item{new_file}{new docx file name as a \code{character}.}
}
\value{
None
}
\description{
Write docx file with original reference.docx and updated style xml
}
\examples{
\dontrun{
  write_styles(xml, "reference.docx", "updated.docx")
}

}
