% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{create_new_style}
\alias{create_new_style}
\title{Create new style from reference style}
\usage{
create_new_style(xml, style_name, ref_style_name = NULL, style_id = "auto")
}
\arguments{
\item{xml}{style xml}

\item{style_name}{new style name.}

\item{ref_style_name}{reference style's name. Default is NULL.}

\item{style_id}{new style id. Unique id in docx is required. By default, it has unique id.}
}
\value{
new style by column?
}
\description{
Create a new style from reference style.
}
