% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_style.R
\name{create_unique_id}
\alias{create_unique_id}
\title{Generate unique style id}
\usage{
create_unique_id(ids)
}
\arguments{
\item{ids}{style id character vector}
}
\value{
character of style id
}
\description{
Generate unique style id
}
