% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_node_by_id}
\alias{get_node_by_id}
\title{To get specific node by style_id}
\usage{
get_node_by_id(styles_xml, node_id)
}
\arguments{
\item{styles_xml}{xml2 node object of styles.}

\item{node_id}{style_id to get specific node.}
}
\value{
xml2 node object
}
\description{
To get specific node by style_id
}
\examples{
\dontrun{
 get_node_by_id(xml, "Author")
}

}
