% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_styles}
\alias{get_styles}
\title{Extract only style tags and its children}
\usage{
get_styles(xml)
}
\arguments{
\item{xml}{\code{xml_node} object with \code{styles} tag as a root. See \code{read_docx}.}
}
\value{
\code{xml_nodeset} object which contains style tags.
}
\description{
Peel off the \code{styles} tag from input \code{xml_node} object. \code{style} tags are contained under the \code{styles} tag.
}
\examples{
\dontrun{
get_styles(xml)
}
}
