% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2df.R
\name{style2df}
\alias{style2df}
\title{Convert \code{docx_styles} object to data.frame}
\usage{
style2df(docx_styles)
}
\arguments{
\item{docx_styles}{docx_styles object.}
}
\value{
data.frame converted from input xml.
}
\description{
Conversion from \code{docx_styles} to data.frame for modification usability.
}
\examples{
\dontrun{
style2df(docx_styles)
}

}
