% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{write_docx}
\alias{write_docx}
\title{Write docx file with original reference.docx and updated style xml}
\usage{
write_docx(styles_xml, org_docx, new_docx)
}
\arguments{
\item{styles_xml}{\code{xml_node} object with \code{styles}}

\item{org_docx}{reference docx file name as a \code{character}. Use files except \code{word/styles.xml}.}

\item{new_docx}{new docx file name as a \code{character}.}
}
\value{
None
}
\description{
Write docx file with original reference.docx and updated style xml
}
\examples{
\dontrun{
  write_docx(xml, "reference.docx", "updated.docx")
}

}
